separate article factory
