+++ src/SWP/Bundle/ContentBundle/DependencyInjection/Configuration.php
@@ -22,7 +22,7 @@ use SWP\Bundle\ContentBundle\Doctrine\ODM\PHPCR\Media;
 use SWP\Bundle\ContentBundle\Doctrine\ORM\Article as ORMArticle;
 use SWP\Bundle\ContentBundle\Doctrine\ORM\ArticleRepository as ORMArticleRepository;
 use SWP\Bundle\ContentBundle\Doctrine\ORM\Route as ORMRoute;
-use SWP\Bundle\ContentBundle\Factory\ArticleFactory;
+use SWP\Bundle\ContentBundle\Factory\PHPCR\ArticleFactory;
 use SWP\Bundle\ContentBundle\Factory\RouteFactory;
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
